requirements.txt
fastapi==0.109.2
llama_cpp_python==0.3.9
uvicorn==0.27.1
whispercppy==0.0.4
python-multipart==0.0.9
opencc-python-reimplemented==0.1.7
pillow==10.4.0
ffmpeg-python==0.2.0
